feat: add fix for the inline-if-can-be-used rule
