+++ docs/releasenotes/9.0.0.md
@@ -367,6 +367,7 @@ to detect metadata without a name.
 TODO
 
 add fix for the deprecated-run-keyword-if rule (#TBD)
+add fix for the inline-if-can-be-used rule (#TBD)
 add fix for the duplicated-variable rule (#1857) (b7a87ed)
 add fix for the else-not-upper-case rule (#1854) (837a979)
 add fix for the empty-return rule (#1846) (f376046)
